+++ src/back/Plants.Application/Commands/EditInstructionCommand.cs
@@ -2,6 +2,6 @@
 
 namespace Plants.Application.Commands;
 
-public record EditInstructionCommand(int InstructionId,  int GroupId, string Text,
+public record EditInstructionCommand(long InstructionId, long GroupId, string Text,
     string Title, string Description, byte[]? CoverImage) : IRequest<EditInstructionResult>;
-public record EditInstructionResult(int InstructionId);
+public record EditInstructionResult(long InstructionId);
