+++ src/back/Plants.Application/Commands/PlaceOrderCommand.cs
@@ -2,5 +2,5 @@
 
 namespace Plants.Application.Commands;
 
-public record PlaceOrderCommand(int PostId, string City, int MailNumber) : IRequest<PlaceOrderResult>;
+public record PlaceOrderCommand(long PostId, string City, int MailNumber) : IRequest<PlaceOrderResult>;
 public record PlaceOrderResult(bool Successfull, string Message);
