+++ src/back/Plants.Application/Commands/StartDeliveryCommand.cs
@@ -2,5 +2,5 @@
 
 namespace Plants.Application.Commands;
 
-public record ConfirmDeliveryCommand(int DeliveryId) : IRequest<ConfirmDeliveryResult>;
+public record ConfirmDeliveryCommand(long DeliveryId) : IRequest<ConfirmDeliveryResult>;
 public record ConfirmDeliveryResult(bool Successfull);
