+++ src/back/Plants.Application/Contracts/IPlantsService.cs
@@ -6,15 +6,15 @@ namespace Plants.Application.Contracts;
 public interface IPlantsService
 {
     Task<IEnumerable<PlantResultItem>> GetNotPosted();
-    Task<PreparedPostResultItem?> GetPrepared(int plantId);
-    Task<CreatePostResult> Post(int plantId, decimal price);
+    Task<PreparedPostResultItem?> GetPrepared(long plantId);
+    Task<CreatePostResult> Post(long plantId, decimal price);
     Task<AddPlantResult> Create(string Name, string Description,
-        int[] Regions, int SoilId,
-        int GroupId, DateTime Created,
+        long[] Regions, long SoilId,
+        long GroupId, DateTime Created,
         byte[][] Pictures);
 
-    Task Edit(int PlantId, string Name, string Description,
-            int[] Regions, int SoilId, 
-            int GroupId, int[] RemovedImages, byte[][] NewImages);
-    Task<PlantResultDto?> GetBy(int id);
+    Task Edit(long PlantId, string Name, string Description,
+            long[] Regions, long SoilId,
+            long GroupId, long[] RemovedImages, byte[][] NewImages);
+    Task<PlantResultDto?> GetBy(long id);
 }
