+++ src/back/Plants.Application/Requests/FindInstructionsRequest.cs
@@ -2,9 +2,9 @@
 
 namespace Plants.Application.Requests;
 
-public record FindInstructionsRequest(int GroupId, string? Title, string? Description) : IRequest<FindInstructionsResult>;
+public record FindInstructionsRequest(long GroupId, string? Title, string? Description) : IRequest<FindInstructionsResult>;
 public record FindInstructionsResult(List<FindInstructionsResultItem> Items);
-public record FindInstructionsResultItem(int Id, string Title, string Description, bool HasCover)
+public record FindInstructionsResultItem(long Id, string Title, string Description, bool HasCover)
 {
     public FindInstructionsResultItem() : this(-1, "", "", false)
     {
