+++ src/back/Plants.Application/Requests/PlantsRequest.cs
@@ -4,7 +4,7 @@ namespace Plants.Application.Requests;
 
 public record PlantsRequest() : IRequest<PlantsResult>;
 public record PlantsResult(List<PlantResultItem> Items);
-public record PlantResultItem(int Id, string PlantName, string Description, bool IsMine)
+public record PlantResultItem(long Id, string PlantName, string Description, bool IsMine)
 {
     //for decoder
     public PlantResultItem() : this(-1, "", "", false)
