+++ src/back/Plants.Infrastructure/Services/PlantsService.cs
@@ -16,7 +16,7 @@ public class PlantsService : IPlantsService
         _ctxFactory = ctxFactory;
     }
 
-    public async Task<PlantResultDto> GetBy(int id)
+    public async Task<PlantResultDto> GetBy(long id)
     {
         var ctx = _ctxFactory.CreateDbContext();
         await using (ctx)
@@ -56,7 +56,7 @@ public class PlantsService : IPlantsService
             }
         }
     }
-    public async Task<PreparedPostResultItem> GetPrepared(int plantId)
+    public async Task<PreparedPostResultItem> GetPrepared(long plantId)
     {
         var ctx = _ctxFactory.CreateDbContext();
         await using (ctx)
@@ -84,7 +84,7 @@ public class PlantsService : IPlantsService
         }
     }
 
-    public async Task<CreatePostResult> Post(int plantId, decimal price)
+    public async Task<CreatePostResult> Post(long plantId, decimal price)
     {
         var ctx = _ctxFactory.CreateDbContext();
         await using (ctx)
@@ -112,8 +112,8 @@ public class PlantsService : IPlantsService
         }
     }
 
-    public async Task<AddPlantResult> Create(string Name, string Description, int[] Regions,
-        int SoilId, int GroupId, DateTime Created, byte[][] Pictures)
+    public async Task<AddPlantResult> Create(string Name, string Description, long[] Regions,
+        long SoilId, long GroupId, DateTime Created, byte[][] Pictures)
     {
         var ctx = _ctxFactory.CreateDbContext();
         await using (ctx)
@@ -132,15 +132,15 @@ public class PlantsService : IPlantsService
                     Created,
                     Pictures
                 };
-                var res = await connection.QueryAsync<int>(sql, p);
+                var res = await connection.QueryAsync<long>(sql, p);
                 var first = res.FirstOrDefault();
                 return new AddPlantResult(first);
             }
         }
     }
 
-    public async Task Edit(int PlantId, string Name, string Description,
-        int[] Regions, int SoilId, int GroupId, int[] RemovedImages, byte[][] NewImages)
+    public async Task Edit(long PlantId, string Name, string Description,
+        long[] Regions, long SoilId, long GroupId, long[] RemovedImages, byte[][] NewImages)
     {
         var ctx = _ctxFactory.CreateDbContext();
         await using (ctx)
